up to 30A then under-voltage fault

+++ src/TMotorCANControl/CAN_manager_servo.py
@@ -495,7 +495,7 @@ class CAN_Manager_servo(object):
         motor_pos= ( pos_int * 0.1)
         motor_spd= (float)( spd_int * 10.0) #motor speed
         motor_cur= (float)( cur_int * 0.01) #motor current
-        motor_temp= data[6]  #motor temperature
+        motor_temp= np.int16(data[6])  #motor temperature
         motor_error= data[7] #motor error mode
         if self.debug:
             print(data)
